Added Lab 5
freeglutapp is the non-OOP template, and glutapp_template is the required template for the Lab 5 submission. Choose to open either one of the folders in VSCode, and in your terminal, type 'make', then './glutapp'. Both have an enemy rectangle and player rectangle. WASD for movement. Holding SPACEBAR shoots a projectile. There is collision detection that detects when the player has slain the enemy and if the player has touched the enemy.