Added Shoot-on-Click, invisible BORDERS, Edited Background
Will now shoot a grey projectile (bullet) when user clicks anywhere on the Scene. I think I'll change it back to spacebar-to-shoot. Photoshop'd background so our character doesn't walk over the trash can and trees and destroy the perspective-feeling. Added INVISIBLE BORDERS that will push the character back and not allow access to that area (off-screen and into the horizon). All that's left to do is implement hit-detection and prompts (renderText). Almost done! Let's get it!