HIT DETECTION IMPLEMENTED
When bullet touches a mob or the princess, that sprite disappears. Now I need to add renderTexts to prompt the user and to announce whether the user has won or lost. I also need to make it so when the character touches the gun and or the bullet on the ground, they disappear and the user can now shoot bullets.